+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## 0.13.3 (tbd)
 
 * Fixed missing support for `favicon-webpack-plugin` in `piral-cli-webpack`
-- Added support for `apiKeys` in `.piralrc` file
+* Added support for `apiKeys` in `.piralrc` file
+* Added support for more flexible entry module resolution to `pilet build`
 * Added `dashboardPath` and `piralChildren` as options for `renderInstance`
 * Updated used chunk hash length for Webpack
 * Replaced `node-sass` with `sass`